Adding init unit testing for state machine

+++ ROBOT_CODE/RobotBase/src/main/java/frc/robot/state/StateChangeResult.java
@@ -3,14 +3,14 @@ package frc.robot.state;
 public class StateChangeResult {
   public String code;
   public String message;
-  public long timestamp;
+  public double timestamp;
   public Object data;
 
-  public StateChangeResult(String code, String message, long timestamp) {
+  public StateChangeResult(String code, String message, double timestamp) {
     this(code, message, timestamp, null);
   }
 
-  public StateChangeResult(String code, String message, long timestamp, Object data) {
+  public StateChangeResult(String code, String message, double timestamp, Object data) {
     this.code = code;
     this.message = message;
     this.timestamp = timestamp;
